+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Please add to the relevant subsections under Unreleased below on every PR where 
 
 ### Fixed
 
+* Prevent crash on `run.summary` for finished runs by @dmitryduev in https://github.com/wandb/wandb/pull/7440
 * Correctly report file upload errors when using wandb-core by @moredatarequired in https://github.com/wandb/wandb/pull/7196
 * Implemented a stricter check for AMD GPU availability by @dmitryduev in https://github.com/wandb/wandb/pull/7322
 * Fixed `run.save()` on Windows by @timoffex in https://github.com/wandb/wandb/pull/7412